backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# OpenAI & LangChain
openai==1.54.0
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.5
langchain-core==0.3.15

# Document Processing
python-docx==1.1.2
openpyxl==3.1.5
pypdf==5.0.1
python-pptx==1.0.2
reportlab==4.2.5

# Calendar/ICS
icalendar==6.0.1

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# CORS
fastapi-cors==0.0.6